feat(#16): support fractional shares + decimal avg_buy_price
Position.Shares int → float64 across model, DTOs, response, and
validatePositionInput; aiPortfolioCtx.Shares follows. GORM AutoMigrate
handles the SQLite column change. shares <= 0 rejection stays, now
accepts decimals (500.5). All calc paths already multiply float64, so
cost basis/P&L/allocation need no changes.

Frontend: PositionModal shares input step=0.0001, avg_buy_price
step=0.01 (parsed as float). HoldingsTable gets formatShares — up to 4
decimals, trailing zeros stripped (replaces formatNum, now unused).

JSON serializes float64 shares as decimal (no scientific notation for
realistic magnitudes).

+++ ai.go
@@ -39,7 +39,7 @@ var errHermesUnavailable = errors.New("hermes tidak tersedia")
 // --- Context built for Hermes (docs/architecture.md §4 shape) ---
 
 type aiPortfolioCtx struct {
-	Shares    int     `json:"shares"`
+	Shares    float64 `json:"shares"`
 	WeightPct float64 `json:"weight_pct"`
 }
 
@@ -87,7 +87,7 @@ func buildAIContext(db *gorm.DB, ticker string) (aiContext, error) {
 
 	var positions []Position
 	_ = db.Find(&positions).Error
-	totalCost, thisCost, shares := 0.0, 0.0, 0
+	totalCost, thisCost, shares := 0.0, 0.0, 0.0
 	for _, p := range positions {
 		c := float64(p.Shares) * p.AvgBuyPrice
 		totalCost += c
